Develop

+++ SRC/emailManager/emails/views.py
@@ -1,8 +1,9 @@
+import json
+import datetime
 import email.errors
 from django.contrib.postgres.search import SearchVector
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import Q
-from django.db.models.fields import json
 from django.http import JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
 from django.views import View
@@ -16,6 +17,10 @@ from emails.models import *
 from django.contrib.postgres.search import TrigramSimilarity
 from django.db.models.functions import Greatest
 from users.models import Signature
+import logging
+
+logger = logging.getLogger(__name__)
+
 
 class AddNewEmail(LoginRequiredMixin, View):
     template_name = 'emails/compose.html'
@@ -94,11 +99,16 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=to)
                         user_email_mapped_receiver.save()
+                        logger.info(msg=f"the email sent to {to} by {request.user}")
+                    else:
+                        logger.error(f"this user with {to} does not exist")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
                 user_email_mapped_sender.save()
                 messages.success(request, 'Your email sent successfully!')
+                logger.info(msg=f"The email sent to {to} by {request.user}")
+
             if cc != '' and bcc == '':
                 email = save_email()
                 email_receiver = EmailReceiver(email=email)
@@ -112,6 +122,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=to)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {to} by {request.user}")
 
                 for cc in cc.split(','):
                     cc = cc.strip()
@@ -122,6 +133,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=cc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {cc} by {request.user}")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
@@ -140,6 +152,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=to)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {to} by {request.user}")
                 for bcc in bcc.split(','):
                     bcc = bcc.strip()
                     if CustomUser.objects.filter(username=bcc).exists():
@@ -149,6 +162,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=bcc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {bcc} by {request.user}")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
@@ -167,6 +181,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=to)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {to} by {request.user}")
                 for cc in cc.split(','):
                     cc = cc.strip()
                     if CustomUser.objects.filter(username=cc).exists():
@@ -176,6 +191,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=cc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {cc} by {request.user}")
                 for bcc in bcc.split(','):
                     bcc = bcc.strip()
                     if CustomUser.objects.filter(username=bcc).exists():
@@ -185,6 +201,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=bcc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {bcc} by {request.user}")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
@@ -198,6 +215,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                            user=request.user)
                 user_email_mapped_sender.save()
                 messages.success(request, 'Your email moved to drafts!!')
+                logger.info(f"The email moved to drafts {request.user}")
             if cc != '' and bcc == '':
                 email = save_email()
                 email_receiver = EmailReceiver(email=email)
@@ -211,6 +229,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=cc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {cc} by {request.user}")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
@@ -229,6 +248,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=bcc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {bcc} by {request.user}")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
@@ -247,7 +267,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=cc)
                         user_email_mapped_receiver.save()
-
+                        logger.info(f"The email sent to {cc} by {request.user}")
                 for bcc in bcc.split(','):
                     bcc = bcc.strip()
                     if CustomUser.objects.filter(username=bcc).exists():
@@ -257,6 +277,7 @@ class AddNewEmail(LoginRequiredMixin, View):
                                                                      place_holder=inbox,
                                                                      user=bcc)
                         user_email_mapped_receiver.save()
+                        logger.info(f"The email sent to {bcc} by {request.user}")
                 user_email_mapped_sender = UserEmailMapped(email=email,
                                                            place_holder=sentbox,
                                                            user=request.user)
@@ -285,7 +306,22 @@ class EmailInboxView(LoginRequiredMixin, View):
     def get(self, request):
         place_holder = EmailPlaceHolders.objects.get(place_holder='inbox')
         user_inbox = UserEmailMapped.objects.filter(user=self.request.user, place_holder=place_holder)
-        # label = EmailPlaceHolders.objects.filter(creator=request.user)
+        for mail in user_inbox:
+
+            # print("*%"*79)
+            # print(filtering)
+            if Filter.objects.filter(user=request.user, subject=mail.email.subject, sender=mail.email.author).exists():
+                filtering = Filter.objects.filter(user=request.user, subject=mail.email.subject,
+                                                  sender=mail.email.author)
+                filtering.update(email=mail.email)
+                for filters in filtering:
+                    # user_inbox.update(place_holder=filters.place_holder)
+                    if mail.email == filters.email:
+                        print("*%" * 79)
+                        print(mail.email == filters.email)
+                        user_inbox = user_inbox.filter(email=filters.email)
+                        user_inbox.update(place_holder=filters.place_holder)
+
         return render(request, self.template_name, {'inbox': user_inbox})
 
 
@@ -324,7 +360,35 @@ class DetailEmailView(LoginRequiredMixin, View):
         except:
             bcc = ''
 
-        return render(request, self.template, {'object': email, "to": to, "cc": cc, "bcc": bcc, "has_bcc": has_bcc})
+class DetailEmailView(LoginRequiredMixin, View):
+            template = 'emails/detail.html'
+
+            def setup(self, request, *args, **kwargs):
+                self.email_instance = get_object_or_404(Email, pk=kwargs['email_id'])
+                return super().setup(request, *args, **kwargs)
+
+            def get(self, request, email_id):
+                email = Email.objects.get(pk=email_id)
+                try:
+                    to = EmailReceiver.objects.get(email=email).to.exists()
+                except:
+                    to = ''
+                try:
+                    cc = EmailReceiver.objects.get(email=email).cc
+                except:
+                    cc = ''
+                has_bcc = False
+                try:
+                    bcc = EmailReceiver.objects.get(email=email).bcc
+                    if bcc.exists():
+                        has_bcc = True
+                except:
+                    bcc = ''
+                inbox = EmailPlaceHolders.objects.get(place_holder="inbox")
+                email_mapp = UserEmailMapped.objects.filter(email=email, place_holder=inbox)
+                email_mapp.update(is_read=True)
+                return render(request, self.template,
+                              {'object': email, "to": to, "cc": cc, "bcc": bcc, "has_bcc": has_bcc})
 
 
 class AddLabelView(LoginRequiredMixin, View):
@@ -342,7 +406,7 @@ class AddLabelView(LoginRequiredMixin, View):
         place_holder = request.POST['place_holder']
         label = EmailPlaceHolders(place_holder=place_holder, creator=request.user)
         label.save()
-
+        logger.info(f"Label {request.POST['place_holder']} created by {request.user}")
         return redirect('email_view')
 
 
@@ -356,6 +420,7 @@ class AddLabelEmail(LoginRequiredMixin, View):
                                                 user=request.user,
                                                 email=email)
         user_email_mapped_new.save()
+        logger.info(f"Email {email} labeled as {label} by {request.user}")
         return redirect('label_view')
 
 
@@ -366,6 +431,13 @@ class EmailTrashView(LoginRequiredMixin, View):
     def get(self, request):
         place_holder = EmailPlaceHolders.objects.get(place_holder='trash')
         user_trash = UserEmailMapped.objects.filter(user=self.request.user, place_holder=place_holder)
+        for emails in user_trash:
+            days = emails.email.created_date + datetime.timedelta(days=30)
+            user_trash_02 = UserEmailMapped.objects.filter(user=self.request.user, place_holder=place_holder,
+                                                           email__created_date__gte=days)
+            if user_trash_02.exists():
+                for email_trash in user_trash_02:
+                    email_trash.delete()
         label = EmailPlaceHolders.objects.filter(creator=request.user)
         return render(request, self.template_name, {'trash': user_trash, "label": label})
 
@@ -378,6 +450,7 @@ class DeleteEmail(LoginRequiredMixin, View):
         inbox = EmailPlaceHolders.objects.get(place_holder='inbox')
         user_email_mapped = UserEmailMapped.objects.filter(place_holder=inbox, user=request.user, email=email)
         user_email_mapped.update(place_holder=trash)
+        logger.info(f"Email {email} moved to trash by {request.user}")
         return redirect('email_trash')
 
 
@@ -405,10 +478,24 @@ class DeleteLabel(LoginRequiredMixin, View):
         label = EmailPlaceHolders.objects.get(pk=label_id)
         emails = UserEmailMapped.objects.filter(place_holder=label)
         inbox = EmailPlaceHolders.objects.get(place_holder='inbox')
-        for email in emails:
-            user_email_mapped = UserEmailMapped.objects.filter(place_holder=label, user=request.user, email=email.email)
-            user_email_mapped.delete()
+        if emails.exists():
+            for email in emails:
+                user_email_mapped = UserEmailMapped.objects.filter(place_holder=label, user=request.user,
+                                                                   email=email.email)
+                user_email_mapped.delete()
+
+                if Filter.objects.filter(user=request.user, email=email.email, place_holder=label).exists():
+                    filtering = Filter.objects.filter(user=request.user, email=email.email, place_holder=label)
+
+                    user_email_mapped_label = UserEmailMapped(email=email.email, user=request.user,
+                                                              place_holder=inbox)
+                    user_email_mapped_label.save()
+                    filtering.delete()
+                    label.delete()
+                    logger.info(f"Label {label} deleted by {request.user}")
+        else:
             label.delete()
+            logger.info(f"Label {label} deleted by {request.user}")
         messages.success(request, "delete was successfully", 'success')
         return redirect('email_view')
 
@@ -429,15 +516,26 @@ class EmailArchiveView(LoginRequiredMixin, View):
     def get(self, request):
         place_holder = EmailPlaceHolders.objects.get(place_holder='archive')
         email = UserEmailMapped.objects.filter(place_holder=place_holder, user=request.user)
+        for mail in email:
+            if Filter.objects.filter(user=request.user, subject=mail.email.subject, sender=mail.email.author).exists():
+                filtering = Filter.objects.filter(user=request.user, subject=mail.email.subject,
+                                                  sender=mail.email.author)
+                filtering.update(email=mail.email)
+                logger.info(f"Email {email} moved to archive by {request.user}")
+                for filters in filtering:
+                    # user_inbox.update(place_holder=filters.place_holder)
+                    if mail.email == filters.email:
+                        print("*%" * 79)
+                        print(mail.email == filters.email)
+                        email = email.filter(email=filters.email)
+                        email.update(place_holder=filters.place_holder)
+
         return render(request, self.template_name, {"email": email})
 
 
 class Search(LoginRequiredMixin, View):
     def get(self, request):
         user_email_mapped = UserEmailMapped.objects.filter(user=request.user)
-        print('^' * 60)
-        print(request.user)
-        print('^' * 60)
         form = SearchBox()
         final_query = Q()
         if 'search' in request.GET:
@@ -454,14 +552,67 @@ class Search(LoginRequiredMixin, View):
                         Q(email__author__username__icontains=cd),
                         Q.OR
                     )
-                    print('#' * 60)
-                    print(final_query)
-                    print(user_email_mapped)
-                    print('#' * 60)
+
                 search = UserEmailMapped.objects.filter(final_query, user=request.user)
+                logger.info(f"User {request.user} searched {cd}")
         return render(request, 'emails/search.html', {'search': search})
 
 
 class Setting(LoginRequiredMixin, View):
     def get(self, request):
         return render(request, 'emails/setting.html')
+
+
+class AddLabelFilter(LoginRequiredMixin, View):
+    def get(self, request):
+        return render(request, 'emails/add_label_filter.html')
+
+    def post(self, request):
+        place_holder = request.POST['place_holder']
+        sender = request.POST['sender']
+        subject = request.POST['subject']
+        label = EmailPlaceHolders(place_holder=place_holder, creator=request.user)
+        label.save()
+        add_filter = Filter(place_holder=label, sender=sender, subject=subject, user=request.user)
+        add_filter.save()
+        logger.info(f"filtering email with subject {subject} moved to label {label} by {request.user}")
+        return render(request, 'emails/setting.html')
+
+
+@login_required(redirect_field_name='login')
+def search_email(request):
+    if request.method == 'POST':
+        cd = json.loads(request.body).get('searchText')
+        user_email_mapped = UserEmailMapped.objects.filter(user=request.user)
+        final_query = Q()
+        for item in user_email_mapped:
+            final_query.add(
+                Q(
+                    email__subject__icontains=cd,
+
+                ) |
+                Q(email__body__icontains=cd) |
+                Q(email__author__username__icontains=cd),
+                Q.OR
+            )
+
+        data = UserEmailMapped.objects.filter(final_query, user=request.user).values()
+        logger.info(f"User {request.user} searched {cd}")
+        # data = emails.values()
+        for email in data:
+            email_object = Email.objects.get(pk=email['email_id'])
+            email['user_id'] = email_object.author.username
+            email['created_date'] = email_object.created_date.date()
+            email['id'] = email_object.subject
+        print('#' * 80)
+        print(len(data))
+        print(data)
+        print("*" * 80)
+        print(list(data))
+        print('#' * 80)
+        return JsonResponse(list(data), safe=False)  # safe let to return a not json response
+
+
+class ChangeTheme(LoginRequiredMixin, View):
+    def get(self, request):
+        return render(request, 'emails/themes.html')
